% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig_usingPreCal_functionalized.R
\name{cal_conceptWeight}
\alias{cal_conceptWeight}
\title{calculate concept weight}
\usage{
cal_conceptWeight(conceptVec, trList)
}
\arguments{
\item{conceptVec}{A vector containing only the element ids of a concept}

\item{trList}{A vector containing all the training gene ids.}
}
\value{
"numberOfIntersection_numberOfUnion"
}
\description{
analyze two gene sets(one training, one concept), return the number of intersected and united genes
}
\examples{
tmp.conceptVec<-1:10
trList.call<-4:12
tmp.data<-cal_conceptWeight(tmp.conceptVec,trList.call) #split the number of intersection/union.
}
