% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig_usingPreCal_functionalized.R
\name{cal_cumNorWeight}
\alias{cal_cumNorWeight}
\title{Calculate cumulative normalized weights}
\usage{
cal_cumNorWeight(weight, posList)
}
\arguments{
\item{weight}{A vector of weights. The name of each value in the vector is it's Entrez gene ID}

\item{posList}{A list of list containing pathways of genes. Each sublist is a pathway (or a concept), containing Entrez gene IDs}
}
\value{
A vector of cumulative normalized weights
}
\description{
calculate the cumulative normalized weights for a sorted gene list with weights and an on-set gene set
}
\examples{
weight.call<-c(0.1,0.2,0.5,0.7,0.9,1)
names(weight.call)<-1:6
pathway.my<-list()
pathway.my[[1]]<-1:10
pathway.my[[2]]<-5:15
cumuNorWeight<-cal_cumNorWeight(weight.call,pathway.my)
}
