% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig_usingPreCal_functionalized.R
\name{cal_weightMatrix}
\alias{cal_weightMatrix}
\title{load concept database from gmt, calculate concept weights, process the file one line at a time}
\usage{
cal_weightMatrix(trList)
}
\arguments{
\item{trList}{A vector containing all the training gene ids.}
}
\value{
A matrix containing two types of concept weights
}
\description{
calculate concept weight between the training gene set and concepts, with/without penalty for over-representing gene
}
\examples{
trList.call<-1:100
concept.info<-cal_weightMatrix(trList.call)
}
