% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig_usingPreCal_functionalized.R
\name{construct_permuList}
\alias{construct_permuList}
\title{contruct permutation list}
\usage{
construct_permuList(weights.cp, posList, nPermu = 1000)
}
\arguments{
\item{weights.cp}{A vector of weights. The name of each value in the vector is it's Entrez gene ID}

\item{posList}{A list of list, containing pathway genes. See example pathway.hallmark and pathway.c2cp}

\item{nPermu}{Number of permutations. Default 1000}
}
\value{
A list of list, containing all the enrichment scores of each number of randomly selected genes.
}
\description{
contruct permutation list to avoid repeated calculations of same number of randomly selected genes
}
\examples{
weight.call<-c(0.1,0.2,0.5,0.7,0.9,1)
names(weight.call)<-1:6
pathway.my<-list()
pathway.my[[1]]<-1:10
pathway.my[[2]]<-5:15
myPermu.call<-construct_permuList(weight.call,pathway.my)
}
