% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig_usingPreCal_functionalized.R
\name{CSEA}
\alias{CSEA}
\title{Calculate CSEA}
\usage{
CSEA(result.uniConSig, posList, nPermu = 1000)
}
\arguments{
\item{result.uniConSig}{The result generated by "cal_uniConSig"}

\item{posList}{A list of list containing pathways of genes. Each sublist is a pathway (or a concept), containing Entrez gene IDs}

\item{nPermu}{The number of permutations. Default 1000. Note that this number should be consistent with pre-calculated permutation list.}
}
\value{
A table of 3 columns, "pathway names","NES","pValue"
}
\description{
transfer the uniConSig result to function "weightedKS_ofCSEA()"
}
\examples{
##For a training list, starting from uniConSig:
trList.call<-seq_len(100)
result<-cal_uniConSig(trList.call)
pathway.my<-list()
pathway.my[[1]]<-c(1,10,100,10000,10001,10003,100033432) #entrez gene ids of a gene set
pathway.my[[2]]<-c(100033413,100033430,100008589,100033417,100033416,100008588)
result.csea<-CSEA(result,pathway.my)
}
