% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig_usingPreCal_functionalized.R
\name{findES}
\alias{findES}
\title{Find the highest absolute value of weights}
\usage{
findES(cumWeight)
}
\arguments{
\item{cumWeight}{A vector of cumulative normalized weights. The name of each value in the vector is it's Entrez gene ID}
}
\value{
A vector of cumulative normalized weights
}
\description{
Find the highest running sum of the weighted random walk k-s test to be the enrichment score (ES)
}
\examples{
#For a "cumuNorWeight" calculated from cal_cumNorWeight:
weight.call<-c(0.1,0.2,0.5,0.7,0.9,1)
names(weight.call)<-1:6
pathway.my<-list()
pathway.my[[1]]<-1:10
pathway.my[[2]]<-5:15
cumuNorWeight<-cal_cumNorWeight(weight.call,pathway.my)
ES<-findES(cumuNorWeight)
}
