% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig_usingPreCal_functionalized.R
\docType{data}
\name{trList.cgc}
\alias{trList.cgc}
\title{Cancer gene sets from Cancer Gene Census}
\format{A list entrez gene ids.}
\usage{
data(trList.cgc)
}
\description{
A list of cancer genes generated by Cancer Gene Census
}
\author{
Xu Chi (original data from Cancer Gene Census) \email{xuc13@pitt.edu}
}
\keyword{datasets}
