% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniConSig_usingPreCal_functionalized.R
\name{weightedKS_ofCSEA}
\alias{weightedKS_ofCSEA}
\title{Calculate weighted random walk k-s test for CSEA}
\usage{
weightedKS_ofCSEA(weights, positiveList, myPermu, nPermu = 1000)
}
\arguments{
\item{weights}{A vector of weights. The name of each value in the vector is it's Entrez gene ID}

\item{positiveList}{A list of list containing pathways of genes. Each sublist is a pathway (or a concept), containing Entrez gene IDs}

\item{myPermu}{A list of list containing permutations of randomly selected genes for each pathway.}

\item{nPermu}{The number of permutations used throughout the calculations}
}
\value{
A table of 3 columns, "pathway names","NES","pValue"
}
\description{
calculate normalized enrichment score based on weighted gene list(weights), on-set gene set(positiveList), and permutation result(myPermu).
}
\examples{
##For a "result" calculated from cal_uniConSig:
weight.call<-c(0.1,0.2,0.5,0.7,0.9,1)
names(weight.call)<-1:6
pathway.my<-list()
pathway.my[[1]]<-1:10
pathway.my[[2]]<-5:15
myPermu.call<-construct_permuList(weight.call,pathway.my)
myNES<-weightedKS_ofCSEA(weight.call,pathway.my,myPermu.call)
}
